backend/authWS/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-jose
bcrypt==4.0.0
passlib[bcrypt]
pydantic[email]
cryptography
urllib3